+++ controlplane/src/core/bufservices/PlatformService.ts
@@ -78,7 +78,6 @@ import { OpenAIGraphql, isValidUrl } from '@wundergraph/cosmo-shared';
 import { DocumentNode, buildASTSchema, parse } from 'graphql';
 import { validate } from 'graphql/validation/index.js';
 import { uid } from 'uid';
-import { eq } from 'drizzle-orm';
 import { GraphApiKeyDTO, GraphApiKeyJwtPayload, PublishedOperationData } from '../../types/index.js';
 import { Composer } from '../composition/composer.js';
 import { buildSchema, composeSubgraphs } from '../composition/composition.js';
@@ -171,6 +170,11 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           name: req.name,
           labelMatchers: req.labelMatchers,
           routingUrl: req.routingUrl,
+        });
+
+        await fedGraphRepo.createGraphCryptoKeyPairs({
+          federatedGraphId: federatedGraph.id,
+          organizationId: authContext.organizationId,
         });
 
         const subgraphs = await subgraphRepo.listByFederatedGraph(req.name, {
@@ -482,6 +486,7 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           return {
             subgraphs: [],
             graphToken: '',
+            graphRequestToken: '',
             response: {
               code: EnumStatusCode.ERR_NOT_FOUND,
               details: `Federated graph '${req.name}' not found`,
@@ -498,9 +503,27 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
 
         const list = await subgraphRepo.listByFederatedGraph(req.name, { published: true });
 
+        const routerRequestToken = await fedRepo.getGraphSignedToken({
+          federatedGraphId: federatedGraph.id,
+          organizationId: authContext.organizationId,
+        });
+
+        if (!routerRequestToken) {
+          return {
+            subgraphs: [],
+            graphToken: '',
+            graphRequestToken: '',
+            response: {
+              code: EnumStatusCode.ERR,
+              details: 'Router Request token not found',
+            },
+          };
+        }
+
         const tokens = await fedRepo.getRouterTokens({
           organizationId: authContext.organizationId,
           federatedGraphId: federatedGraph.id,
+          limit: 1,
         });
 
         let graphToken: GraphApiKeyDTO;
@@ -544,6 +567,7 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
             labels: g.labels,
           })),
           graphToken: graphToken.token,
+          graphRequestToken: routerRequestToken,
           response: {
             code: EnumStatusCode.OK,
           },
@@ -1160,6 +1184,7 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
         const authContext = await opts.authenticator.authenticate(ctx.requestHeader);
         const fedGraphRepo = new FederatedGraphRepository(opts.db, authContext.organizationId);
         const subgraphRepo = new SubgraphRepository(opts.db, authContext.organizationId);
+        const schemaCheckRepo = new SchemaCheckRepository(opts.db);
 
         const graph = await fedGraphRepo.byName(req.graphName);
 
@@ -1169,7 +1194,10 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
               code: EnumStatusCode.ERR_NOT_FOUND,
               details: 'Requested graph does not exist',
             },
+            compositionErrors: [],
+            changes: [],
             affectedGraphs: [],
+            trafficCheckDays: 0,
           };
         }
 
@@ -1182,22 +1210,14 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
               code: EnumStatusCode.ERR_NOT_FOUND,
               details: 'Requested check not found',
             },
+            compositionErrors: [],
+            changes: [],
             affectedGraphs: [],
+            trafficCheckDays: 0,
           };
         }
 
-        let addCount = 0;
-        let minusCount = 0;
-        for (const log of checkDetails.changes) {
-          if (log.changeType.includes('REMOVED')) {
-            minusCount += 1;
-          } else if (log.changeType.includes('ADDED')) {
-            addCount += 1;
-          } else if (log.changeType.includes('CHANGED')) {
-            addCount += 1;
-            minusCount += 1;
-          }
-        }
+        const { trafficCheckDays } = await schemaCheckRepo.getFederatedGraphConfigForCheckId(req.checkId, graph.id);
 
         return {
           response: {
@@ -1206,10 +1226,9 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           check,
           affectedGraphs: check.affectedGraphs,
           proposedSubgraphSchemaSDL: check.proposedSubgraphSchemaSDL,
-          changeCounts: {
-            additions: addCount,
-            deletions: minusCount,
-          },
+          changes: checkDetails.changes,
+          compositionErrors: checkDetails.compositionErrors,
+          trafficCheckDays,
         };
       });
     },
@@ -3547,6 +3566,7 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
         const tokens = await fedRepo.getRouterTokens({
           organizationId: authContext.organizationId,
           federatedGraphId: federatedGraph.id,
+          limit: 100,
         });
 
         return {