+++ router/cmd/instance.go
@@ -2,13 +2,14 @@ package cmd
 
 import (
 	"fmt"
+	"github.com/wundergraph/cosmo/router/internal/controlplane/configpoller"
+	"github.com/wundergraph/cosmo/router/internal/controlplane/selfregister"
 	"go.uber.org/automaxprocs/maxprocs"
 
 	"github.com/wundergraph/cosmo/router/authentication"
 	"github.com/wundergraph/cosmo/router/config"
 	"github.com/wundergraph/cosmo/router/core"
 	nodev1 "github.com/wundergraph/cosmo/router/gen/proto/wg/cosmo/node/v1"
-	"github.com/wundergraph/cosmo/router/internal/controlplane"
 	"github.com/wundergraph/cosmo/router/internal/handler/cors"
 	"github.com/wundergraph/cosmo/router/internal/metric"
 	"github.com/wundergraph/cosmo/router/internal/trace"
@@ -28,7 +29,8 @@ func NewRouter(params Params) (*core.Router, error) {
 	}
 
 	var routerConfig *nodev1.RouterConfig
-	var cp controlplane.ConfigFetcher
+	var configPoller configpoller.ConfigPoller
+	var selfRegister selfregister.SelfRegister
 
 	cfg := params.Config
 	logger := params.Logger
@@ -38,35 +40,16 @@ func NewRouter(params Params) (*core.Router, error) {
 		if err != nil {
 			logger.Fatal("Could not read router config", zap.Error(err), zap.String("path", cfg.RouterConfigPath))
 		}
-
-		if cfg.Graph.Token == "" {
-			logger.Warn("Static router config file provided, but no graph token. Disabling schema usage tracking, thus breaking change detection. Not recommended for production use.")
-			cfg.GraphqlMetrics.Enabled = false
-
-			// Only disable default tracing and metrics if no custom OTLP exporter is configured
-			if cfg.Telemetry.Tracing.Enabled && len(cfg.Telemetry.Tracing.Exporters) == 0 {
-				cfg.Telemetry.Tracing.Enabled = false
-			}
-			if cfg.Telemetry.Metrics.OTLP.Enabled && len(cfg.Telemetry.Metrics.OTLP.Exporters) == 0 {
-				cfg.Telemetry.Metrics.OTLP.Enabled = false
-			}
-
-			// Show warning when no custom OTLP exporter is configured and default tracing/metrics are disabled
-			// due to missing graph token
-			if !cfg.Telemetry.Tracing.Enabled && len(cfg.Telemetry.Tracing.Exporters) == 0 {
-				logger.Warn("Static router config file provided, but no graph token. Disabling default tracing. Not recommended for production use.")
-			}
-			if !cfg.Telemetry.Metrics.OTLP.Enabled && len(cfg.Telemetry.Metrics.OTLP.Exporters) == 0 {
-				logger.Warn("Static router config file provided, but no graph token. Disabling default OTLP metrics. Not recommended for production use.")
-			}
-		}
 	} else {
-		cp = controlplane.New(
-			controlplane.WithControlPlaneEndpoint(cfg.ControlplaneURL),
-			controlplane.WithFederatedGraph(cfg.Graph.Name),
-			controlplane.WithLogger(logger),
-			controlplane.WithGraphApiToken(cfg.Graph.Token),
-			controlplane.WithPollInterval(cfg.PollInterval),
+		configPoller = configpoller.New(cfg.Graph.Name, cfg.ControlplaneURL, cfg.Graph.Token,
+			configpoller.WithLogger(logger),
+			configpoller.WithPollInterval(cfg.PollInterval),
+		)
+	}
+
+	if cfg.RouterRegistration {
+		selfRegister = selfregister.New(cfg.ControlplaneURL, cfg.Graph.Token,
+			selfregister.WithLogger(logger),
 		)
 	}
 
@@ -97,7 +80,8 @@ func NewRouter(params Params) (*core.Router, error) {
 		core.WithListenerAddr(cfg.ListenAddr),
 		core.WithOverrideRoutingURL(cfg.OverrideRoutingURL),
 		core.WithLogger(logger),
-		core.WithConfigFetcher(cp),
+		core.WithConfigPoller(configPoller),
+		core.WithSelfRegistration(selfRegister),
 		core.WithIntrospection(cfg.IntrospectionEnabled),
 		core.WithPlayground(cfg.PlaygroundEnabled),
 		core.WithGraphApiToken(cfg.Graph.Token),
@@ -136,6 +120,7 @@ func NewRouter(params Params) (*core.Router, error) {
 			AllowHeaders:     cfg.CORS.AllowHeaders,
 			MaxAge:           cfg.CORS.MaxAge,
 		}),
+		core.WithDevelopmentMode(cfg.DevelopmentMode),
 		core.WithTracing(traceConfig(&cfg.Telemetry)),
 		core.WithMetrics(metricsConfig(&cfg.Telemetry)),
 		core.WithEngineExecutionConfig(cfg.EngineExecutionConfiguration),
@@ -149,6 +134,7 @@ func traceConfig(cfg *config.Telemetry) *trace.Config {
 	var exporters []*trace.Exporter
 	for _, exp := range cfg.Tracing.Exporters {
 		exporters = append(exporters, &trace.Exporter{
+			Disabled:      exp.Disabled,
 			Endpoint:      exp.Endpoint,
 			Exporter:      exp.Exporter,
 			BatchTimeout:  exp.BatchTimeout,
@@ -169,6 +155,7 @@ func metricsConfig(cfg *config.Telemetry) *metric.Config {
 	var openTelemetryExporters []*metric.OpenTelemetryExporter
 	for _, exp := range cfg.Metrics.OTLP.Exporters {
 		openTelemetryExporters = append(openTelemetryExporters, &metric.OpenTelemetryExporter{
+			Disabled: exp.Disabled,
 			Endpoint: exp.Endpoint,
 			Exporter: exp.Exporter,
 			Headers:  exp.Headers,