+++ studio/src/components/reactflow-graph-node.tsx
@@ -8,7 +8,7 @@ function ReactFlowGraphNode({ data }: Node) {
       {data.parentId && (
         <Handle type="target" position={Position.Top} isConnectable={false} />
       )}
-      <div className="nodrag dark:ring-white/15 flex w-[110px] items-center justify-center gap-x-1 rounded border border-transparent bg-secondary px-1 py-1 text-left text-xs shadow-sm ring-1 ring-black/[.08] transition duration-150 dark:shadow-black/60">
+      <div className="nodrag dark:ring-white/15 border-border-emphasized flex w-[110px] items-center justify-center gap-x-1 rounded border bg-white px-1 py-1 text-left text-xs shadow-sm shadow-black/5 ring-1 ring-black/[.08] transition duration-150 dark:bg-secondary dark:shadow-black/60">
         <div className="flex items-center justify-center rounded-full p-1">
           {data.kind === "graph" ? (
             <PiGraphLight className="h-3 w-3 text-secondary-foreground" />