+++ studio/src/hooks/use-user.ts
@@ -1,0 +1,6 @@
+import { UserContext } from "@/components/app-provider"
+import { useContext } from "react"
+
+export const useUser = () => {
+    return useContext(UserContext)
+}