+++ connect/CHANGELOG.md
@@ -4,6 +4,12 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.34.0](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-connect@0.33.0...@wundergraph/cosmo-connect@0.34.0) (2023-11-30)
+
+### Features
+
+* register router on the controlplane ([#318](https://github.com/wundergraph/cosmo/issues/318)) ([10f86df](https://github.com/wundergraph/cosmo/commit/10f86dfebd80265d42015eaf3b9c15f941aef66b)) (@StarpTech)
+
 # [0.33.0](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-connect@0.32.0...@wundergraph/cosmo-connect@0.33.0) (2023-11-29)
 
 ### Bug Fixes